// cms/src/payload-types.ts
export interface Pagina {
  id: number;
  titulo: string;
  /**
   * Gerado automaticamente a partir de "titulo". Edite apenas se necessário.
   */
  slug?: string | null;
  menuGrupo?:
    | ('' | 'institucional' | 'integrantes' | 'atividade' | 'legislacao' | 'transparencia' | 'atendimento')
    | null;
  /**
   * Se vazio, usa o título da página.
   */
  menuLabel?: string | null;
  menuDesc?: string | null;
  menuIcone?:
    | (
        | 'documento'
        | 'arquivo'
        | 'calendario'
        | 'pessoas'
        | 'lista'
        | 'predio'
        | 'predio2'
        | 'livro'
        | 'megafone'
        | 'balao'
      )
    | null;
  conteudo?: {
    root: {
      type: string;
      children: {
        type: any;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  updatedAt: string;
  createdAt: string;
}